[Refactor] refactoring formatter
